Tardis/requirements.txt
alembic==1.14.0
blinker==1.9.0
cachelib==0.9.0
click==8.1.7
dominate==2.9.1
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-DebugToolbar==0.16.0
Flask-Env==2.0.0
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.7
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-flask==1.3.0
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
